engine/requirements-django3.9.txt
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
# backports.zoneinfo==0.2.1
beautiful-date==2.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
constantly==15.1.0
crispy-bootstrap4==2022.1
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==4.2
django-allauth==0.54.0
django-appconf==1.0.5
django-cachalot==2.6.1
django-ckeditor==6.7.0
django-ckeditor-5==0.2.8
django-crispy-forms==2.0
django-cryptography==1.1
django-debug-toolbar==4.1.0
django-etc==1.4.0
django-hitcount==1.3.5
django-import-export==3.2.0
django-js-asset==2.1.0
django-mobile-detect==0.1
django-money==3.2.0
django-multiselectfield==0.1.12
django-outbox-hitcount==1.0.11
django-outbox-menu==1.0.31
django-outbox-region==1.0.2
django-parler==2.3
django-recaptcha==3.0.0
django-uuslug==2.0.0
et-xmlfile==1.1.0
filelock==3.12.2
gcsa==2.1.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.60.0
httpagentparser==1.9.5
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonfield==3.1.0
MarkupPy==1.14
midtransclient==1.4.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
outbox-encryption==1.1.3
pbr==5.11.1
Pillow==9.5.0
platformdirs==3.10.0
protobuf==4.23.4
py-moneyed==3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.10.0
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stevedore==5.1.0
tablib==3.5.0
text-unidecode==1.3
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.2
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
